requirements.txt
django
pillow
django-allauth
django-crispy-forms
crispy-bootstrap5
